change time type from duration to int, fix small bugs

+++ src/main/java/org/demo/codesharingplatform/dtos/mapper/CodeMapper.java
@@ -15,7 +15,7 @@ public class CodeMapper {
             return new CodeDTO(
                     code.getDate().toString(),
                     code.getCode(),
-                    (int) code.getSecondsLeft(),
+                    code.getSecondsLeft(),
                     code.getViews()
             );
         } else {
